+++ NEWS.md
@@ -1,3 +1,12 @@
+# GSNA 0.1.6.0
+* Development version
+* Added the `termSummary()` function (along with some associated helper functions `filterTerms()` and `score_word_clusters()`) that can take multiple gene set names/terms and convert them to summarized names.    
+* Added the `gsnSubnetsDotPlot()` function to provide an additional way to represent gene set subnets, ploting gene sets grouped on the y-axis by subnet, positioned on the x-axis by a configurable statistical column, with scaling by number of gene set members and coloring by an additional configural statistic.  
+* Added the `prepend_subnet` function for the `gsnPlotNetwork()` function.  
+
+# GSNA 0.1.5.2
+* Development version  
+* The `prepend_subnet` argument was added to the gsnPlotNetwork function, allowing the users to prepend the subnet name to gene set identifiers when plotting networks. This feature is now enabled by default.
 
 # GSNA 0.1.5.1
 * Development version